Don't expose OpenGL API to application.
I'm not sure what this was intended for and minetest doesn't
seem to use it, but it prevents us from building for
Direct3D.

Maybe minetest will use it in the future? We'll cross that
bridge when we come to it.

This reverts commit 5bf68b5731b1817c31959b8e3adf19a4c2307630.
This reverts commit f83dab83cf691014f5953a8418b799cfcc0c4301.
This reverts commit da22e5d7c04bccc9bbe86db0a4e7c8f9cf5f22e5.

See https://github.com/minetest/irrlicht/pull/52 for where
this was originally introduced.